# packages/api-contract/openapi.yaml
openapi: 3.1.0
info:
  title: Band Manager API
  version: 0.1.0
servers:
  - url: http://localhost:8080
paths:
...
  /inventory:
    get:
      operationId: listInventory
      summary: List active inventory products and variants
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Inventory list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InventoryResponse'
        '401':
          description: Missing or invalid session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    InventoryResponse:
      type: object
      required:
        - products
      properties:
        products:
          type: array
          items:
            $ref: '#/components/schemas/InventoryProduct'
...
    InventoryProduct:
      type: object
      required:
        - id
        - bandId
        - name
        - category
        - photo
        - variants
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
        bandId:
          type: string
          format: uuid
        name:
          type: string
        category:
          $ref: '#/components/schemas/InventoryCategory'
        photo:
          $ref: '#/components/schemas/InventoryPhoto'
        variants:
          type: array
          items:
            $ref: '#/components/schemas/InventoryVariant'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    InventoryVariant:
      type: object
      required:
        - id
        - productId
        - size
        - colour
        - price
        - cost
        - quantity
        - soldOut
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
        productId:
          type: string
          format: uuid
        size:
          $ref: '#/components/schemas/InventorySize'
        colour:
          type: string
        price:
          $ref: '#/components/schemas/Money'
        cost:
          $ref: '#/components/schemas/Money'
        quantity:
          type: integer
          minimum: 0
        soldOut:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    InventoryPhoto:
      type: object
      required:
        - objectKey
        - contentType
        - sizeBytes
      properties:
        objectKey:
          type: string
          minLength: 1
        contentType:
          type: string
          minLength: 1
        sizeBytes:
          type: integer
          minimum: 1
    Money:
      type: object
      required:
        - amount
        - currency
      properties:
        amount:
          type: integer
          minimum: 0
        currency:
          type: string
          enum:
            - BRL
...
    InventoryCategory:
      type: string
      enum:
        - shirt
        - hoodie
        - tote_bag
        - patch
        - sticker
        - vinyl
        - cd
        - cassette
        - accessory
    InventorySize:
      type: string
      enum:
        - not_applicable
        - one_size
        - pp
        - p
        - m
        - g
        - gg
        - xgg
...
    ErrorResponse:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
        message:
          type: string